* Fix gldebug warning for sign conversion (cast to silence) * Fix bug on MSW where export pos routines in dialog could fire before dialog was visible

+++ src/gl/glDebug.h
@@ -32,7 +32,7 @@
 #define glError() { \
 		GLenum err = glGetError(); \
 		while (err != GL_NO_ERROR) { \
-					fprintf(stderr, "glError: %s caught at %s:%u\n", (char *)gluErrorString(err), __FILE__, __LINE__); \
+					fprintf(stderr, "glError: %s caught at %s:%u\n", (char *)gluErrorString(err), __FILE__, (unsigned int)__LINE__); \
 					err = glGetError(); \
 				} \
 		}
